msm_fb: Added hardware workaround for MDP 3.1
The source image which is drawn on the
famebuffer memory before sending to a display panel
is broken into 2 or 4 smaller regions
if its width and height on the frame buffer
(destination) meets any of the following conditions:

1. Source image splits in case it is drawn on
framebuffer memory with destination width mod 32 is
14,22,30 or 6 and destination format is 32 bpp.

2. Source image splits in case destination height
or width mod 32 is 3 or 1.

A special case is added in above cases for destination width
or height of 3 to avoid breaking of source window to width
and height of 0 respectively.

Scaling of source image to destination size is changed to use
MN scaling option from hardware in case source size is more than
4 * destination size - 4.

Removed unused arguments from mdp_ppp_blit function.

Change-Id: Ibb51a5cbd3cd6bc5cbfb87af9737f9d6fa08b354
Signed-off-by: Neeraj Kumar <c_neekum@quicinc.com>