msm: clock-7x30: Fix power rail enabling order
Power rails should be turned on after clocks are enabled and
turned off before clocks are disabled. It must be done in this
order since the modem needs the clocks to be on to properly turn
on and off the power rails.

It just so happens that sometimes the modem enables and disables
the clock branch when the clock's respective power rail is
enabled and disabled. Changing the ordering shouldn't be a
problem since the worst that could happen is the enable bits in
the clock's register are written twice (once by _soc_clk_enable()
and once by the modem's power rail code via proc_comm).

Change-Id: I987f83e8f46545e406e9ae386cf73daca7f0ba0e
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>