msm: clock-7x30: Pair camif_pad_pclk ownership with cam_m_clk
The ownership for the camif_pad_pclk is determined by the global
ownership register while the ownership for the cam_m_clk is
determined by the apps3 register. Controlling the camif_pad_pclk
locally while not controlling the cam_m_clk can lead to
dependency issues and problems. Therefore only allow these two
clocks to be locally controlled if the cam_m_clk is locally
owned.

Change-Id: Ieaf687b00c207e372f2583d33be7332125bab408
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>